Point of Sale System
This system emulates a process one would find at cash registers also known as Point of Sale systems. The program intakes 4 items at any price and calculates the amount due as well as the average price of the items bought. Additionally, the program takes in the amount tendered and outputs the change due. Finally, the program calculates the number of dollars, quarters, dimes, nickels, and pennies that are owed, making sure to use the highest denomination of change first.